📝 feat. payment 페이지 컴포 폴더명 및 경로변경
NewpayInput, OrderPdtInfo, PayAgree, PayMethod, PayMethodInput
