+++ src/components/paymentPage/newPayInput/NewPayInput.style.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const DivArea = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+`
+
+export const InputDivArea = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  margin-top: 10px;
+`
