+++ src/components/paymentPage/orderPdtInfo/OrderPdtInfo.tsx
@@ -1,0 +1,9 @@
+import * as t from './OrderPdtInfo.style'
+
+export const FreeShipping = () => {
+  return (
+    <>
+      <t.ShippingFee>배송비 0원</t.ShippingFee>
+    </>
+  )
+}
